PERSONAL EXPENSE TRACKER
PERSONAL EXPENSE TRACKER
For you to be able to run this code from the main.py, you have to make sure you have installed python3.9 or python3.13 and also installed this python packages (matplotlib and NumPy)
The Pie_Chart.py can not be run by the main.py because it will give errors. I tried to fix the problem but non of them worked, but the Pie_chart.py does work on it's own(by going inside the file and pressing the run button.).